requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
torch
PyMuPDF
python-dotenv
